Groups: Cache calls to BP_Groups_Member::get_group_administrator_ids()
When a logged-in user is viewing the group directory and is a member of
each group in the loop, when rendering the "Leave Group" button, a call
is done to fetch the group admins of each group to determine whether the
user is the sole admin.

Since this query is uncached, this could lead to twenty additional DB
queries being run on the page.  Thus, this commit caches this query and
invalidates the cache whenever a group member is promoted / demoted or
saved into the database.

Fixes #5711.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8537 cdf35c40-ae34-48e0-9cc9-0c9da1808c22