Welcome screen: add strings for maintenance/security release notes
These have been copied from WordPress core, which uses them to update
their welcome screen for minor releases without causing string changes
(or tries to).

Since the layout of our welcome screen is not that dissimilar from what
WordPress’, we should see if we are able to make use of these in any
future 2.1.x minor releases.

If we can, this will help our welcome screen be more relevant and
informative.

See #5805

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8779 cdf35c40-ae34-48e0-9cc9-0c9da1808c22