In {{{bp_blogs_add_user_to_blog ()}}} function make sure getting user roles returns an array
On multisite configs, when a member registers, he is not assigned to any blog  and has no role.
When activating the user from the Manage signups screen, getting his role in {{{bp_blogs_add_user_to_blog()}}} was returning an empty string causing a notice error when trying to get the keys of the "capabilities" usermeta.
To avoid this, we now make sure this usermeta is an array.

Fixes #5834

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8965 cdf35c40-ae34-48e0-9cc9-0c9da1808c22