bp-legacy: Properly clear cookies when logging out.
When a user logs out, buddypress.js attempts to clear all cookies. However,
this now fails with BP 2.1.

The problem is when we enqueued our JS assets (#5208), we upgraded
jquery.cookie to 1.4.1.  v1.4.1 uses a new routine to clear cookies.

This commit updates our clear cookies calls to this updated routine.

Fixes #5909 (2.1-branch).

git-svn-id: https://buddypress.svn.wordpress.org/branches/2.1@9055 cdf35c40-ae34-48e0-9cc9-0c9da1808c22