Make sure themes can override BuddyPress stylesheet.
Since version 2.1, BuddyPress minified stylesheet is used if SCRIPT_DEBUG is not set. This was preventing themes to override the stylesheet using a file named buddypress.css. This patch make sure that if a buddypress.css file is located in one of the allowed css subfolder it will override the one provided by BuddyPress.

props r-a-y, shpitzyl

See #5888 (branch 2.1)

git-svn-id: https://buddypress.svn.wordpress.org/branches/2.1@9064 cdf35c40-ae34-48e0-9cc9-0c9da1808c22