travis-ci: specify IRC server/channel for notifications via encrypted key instead of server env var.
travis-ci.org support has told us that notifications are not sent from the build VM which runs our test script (which is where the env variables are set), and are instead from a seperate component which doesn't take the env vars into account. This is why the previous two attempts on r9080 and r9081 did not work.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@9083 cdf35c40-ae34-48e0-9cc9-0c9da1808c22