Make sure Site admins can set the "register" and "activate" pages in the pages association settings screen even if signups are not allowed
In 2.2, when signups are not allowed, it is not possible to associate a WordPress page to the register or the activate specific components. It introduced some confusion for some site admins and had an impact on plugins allowing users to invite others.

We will keep on working on the best way to deal with these directory pages in the next release. In the meantime, we are bringing back the possibility to always set the association for the register and the activate specific components.

props boonebgorges

See #6176
Fixes #6197 (branch 2.2)

git-svn-id: https://buddypress.svn.wordpress.org/branches/2.2@9479 cdf35c40-ae34-48e0-9cc9-0c9da1808c22