Rename component admin files to avoid confusion with non-admin files sharing the same name.
This rename de-duplicates the file names and make the context they're loaded in clearer.
This is a change for developer convenience. e.g. some IDEs will offer a shortcut to opening the files by name, but is very confusing when two identically-named files are listed in the results with, usually, no quick way of disambiguating them.

Fixes #6051


git-svn-id: https://buddypress.svn.wordpress.org/trunk@9503 cdf35c40-ae34-48e0-9cc9-0c9da1808c22