Don't create a current user by default in most unit tests.
The default state of our tests should be as a logged-out user. BP does pretty
different things with different kinds of logged-in users, so individual tests
should decide for themselves what the current user status should be.

Moreover, the creation of fixture users is fairly resource-intensive. Not
creating users when they're not needed makes the test suite run 5-10% faster.

See #6009.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@9561 cdf35c40-ae34-48e0-9cc9-0c9da1808c22