Enforce no_results when getting a 0=1 result from the member type tax query.
When passing a 'member_type' to `BP_User_Query` and no users belong to the
member type, the `WP_Tax_Query` transformation will return a '0 = 1' clause to
indicate that no results were found. We should respect this result, so that no
members are returned for this kind of query.

Fixes #6334 for the 2.2 branch.

git-svn-id: https://buddypress.svn.wordpress.org/branches/2.2@9663 cdf35c40-ae34-48e0-9cc9-0c9da1808c22