Don't trailingslash the href of subnav items if the 'link' was provided explicitly in `bp_core_new_subnav_item()`.
`trailingslashit()` is not smart enough to detect links with query params, so
that links can be improperly trailingslashed in some cases. More generally,
it's not the case that all links should be trailingslashed (in particular,
those that do not lead to WP/BP pages), so plugin authors should be responsible
for trailingslashing on their own.

Props pareshradadiya.
See #6353.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@9701 cdf35c40-ae34-48e0-9cc9-0c9da1808c22