Avatar UI: Add Conditional functions and new BP Attachment methods
- `bp_avatar_is_front_edit()` checks if the user is editing his or a group avatar on front-end
- `bp_avatar_use_webcam()` checks if the Camera feature should be enabled. By default, it is not the case for touch devices as it can introduce some confustions with the possibility of using the camera of the device to take a selfie, save it as an image an upload it using the Browse button of the Avatar UI. Chrome and Firefox are the only browsers supporting getUserMedia.

NB: these two functions can be filtered to disallow the Camera feature or to completely disallow the new Avatar UI if you prefer to carry on using the legacy UI.

`BP_Attachment::script_data()` and `BP_Attachment_Avatar::script_data()` are two new methods to build the javascript localization data. The `bp_attachment_avatar_params` filter is fired if the current object is not a group or a user and can be used to build the params of another object such as a blog. You can also use the filter `bp_attachment_avatar_script_data` if you need to override data for any object.

Props boonebgorges, DJPaul.

See #6290
See #6278

git-svn-id: https://buddypress.svn.wordpress.org/trunk@9755 cdf35c40-ae34-48e0-9cc9-0c9da1808c22