Improve `bp_core_get_allowed_avatar_types()`. Make sure to get the file type if the upload was not done using html5
The way we used to check the image file type is not working if the browser is Internet Explorer < 10. In this particular case, the Plupload runtime is falling back to flash and the file type is `application/octet-stream`. We are now using `wp_check_filetype_and_ext()` to fix this issue.

You can use the filter `bp_core_get_allowed_avatar_types` if you wish to *restrict* the avatar image types. Our supported types are: jpg, jpeg, png and gif.

Props DJPaul.

Fixes #6336
See #6290
See #6278

git-svn-id: https://buddypress.svn.wordpress.org/trunk@9760 cdf35c40-ae34-48e0-9cc9-0c9da1808c22