BP User Query: Remove invalid IDs from `$user_ids` property after query is run.
If the `user_ids` parameter is passed to `BP_User_Query` with invalid user
IDs, after the user query is run, the `user_ids` property still references
these invalid user IDs.  This can lead to devs to assume that the user ID
exists when it does not.

This commit removes these invalid user IDs and includes a unit test.

See #6577.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10043 cdf35c40-ae34-48e0-9cc9-0c9da1808c22