Introduce `fields` parameter to `bp_has_activities()` stack.
`fields=ids` will fetch only the IDs of matched activities. `fields=all`
(or any other value of `fields`, for the moment) will return full activity
objects.

The structure of the array returned from `BP_Activity_Activity::get()` is the
same regardless of the value of `fields`: matched items are stored in the
'activities' member of the returned array.

Props r-a-y.
Fixes #6426.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10217 cdf35c40-ae34-48e0-9cc9-0c9da1808c22