Messages: Return context-appropriate error messages for `messages_new_message()`.
This commit introduces the `error_type` parameter in the
`messages_new_message()` function, which allows for us to return
descriptive error messages to the user when `'error_type'` is set to
`'wp_error'`.

This most notably addresses an issue when entering an invalid username
when composing a new private message.

Commit also includes a unit test.

Fixes #6535.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10286 cdf35c40-ae34-48e0-9cc9-0c9da1808c22