Specify richtext xprofile field editor height on a per-context basis.
When displaying the edit format preview on Dashboard > Users > Profile Fields,
the default 'textarea_rows' value of 20 was far too high. This changeset
changes it to 1, and specifies a value of 10 for other instances of the editor.

To accommodate these different values, a `$context` parameter is added to the
`bp_xprofile_field_type_textarea_editor_args` filter, with possible values
'edit' and 'admin'.

Props imath.
Fixes #6722.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10328 cdf35c40-ae34-48e0-9cc9-0c9da1808c22